feat: DAH-2661 handle code complexity
